refactor: move _is_schedule_name_conflict into each backend's shared utils
Three distinct matchers (sqlite string match, postgres sqlstate, mongo
keyPattern) stay distinct; the sync/async twin copies collapse into the
family's existing cross-twin home so they cannot drift.
