+++ libs/agno/agno/db/postgres/postgres.py
@@ -27,6 +27,7 @@ from agno.db.base import (
 from agno.db.migrations.manager import MigrationManager
 from agno.db.postgres.schemas import get_table_schema_definition
 from agno.db.postgres.utils import (
+    _is_schedule_name_conflict,
     apply_sorting,
     bulk_upsert_metrics,
     calculate_date_metrics,
@@ -134,29 +135,6 @@ def _shared_table_init_lock(database_key: str) -> RLock:
     """Return one in-process lazy-DDL lock for every database schema."""
     with _TABLE_INIT_LOCKS_GUARD:
         return _TABLE_INIT_LOCKS.setdefault(database_key, RLock())
-
-
-def _is_schedule_name_conflict(error: IntegrityError, table_name: str, schema_name: str) -> bool:
-    """Match only the generic or actor-scoped schedule-name indexes."""
-    original = error.orig
-    cause = getattr(original, "__cause__", None)
-    sqlstate = (
-        getattr(original, "sqlstate", None) or getattr(original, "pgcode", None) or getattr(cause, "sqlstate", None)
-    )
-    diagnostic = getattr(original, "diag", None)
-    constraint_name = getattr(diagnostic, "constraint_name", None) or getattr(cause, "constraint_name", None)
-    reported_table = getattr(diagnostic, "table_name", None) or getattr(cause, "table_name", None)
-    reported_schema = getattr(diagnostic, "schema_name", None) or getattr(cause, "schema_name", None)
-    return (
-        sqlstate == "23505"
-        and constraint_name
-        in {
-            f"{table_name}_uq_generic_name",
-            f"{table_name}_uq_studio_owner_name",
-        }
-        and reported_table == table_name
-        and reported_schema == schema_name
-    )
 
 
 class PostgresDb(BaseDb):
